# swagger-docs/gateway-openapi.yaml
openapi: 3.0.3
info:
  title: Gateway Service
  version: "1.0.0"
servers:
  - url: http://localhost:8080
    description: Gateway local
  - url: http://gateway-srv:8080
    description: Gateway in-cluster
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

    apiKeyAuth:
      type: apiKey
      in: header
      name: X-Api-Key
      description: >
        Developer API key. Only accepted on /v1/checks and /checks.
        Gateway exchanges it internally for a short-lived JWT via Auth Service.
...
paths:
...
  /auth/admin/users:
    get:
      summary: List users (admin)
      tags: [Auth Admin]
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: is_admin
          schema:
            type: boolean
        - in: query
          name: access_status
          schema:
            type: string
            enum: [approved, pending, rejected]
        - in: query
          name: sort
          schema:
            type: string
            enum: [created_at, user_name, email]
        - in: query
          name: order
          schema:
            type: string
            enum: [asc, desc]
        - in: query
          name: page
          schema:
            type: integer
        - in: query
          name: page_size
          schema:
            type: integer
      responses:
        "200":
          description: Paginated user list
        "401":
          description: Unauthorized
        "403":
          description: Admin privileges required
    post:
      summary: Create user (admin)
      tags: [Auth Admin]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [user_name, email, date_of_birth]
              properties:
                user_name:
                  type: string
                email:
                  type: string
                  format: email
                password:
                  type: string
                is_admin:
                  type: boolean
                date_of_birth:
                  type: string
      responses:
        "201":
          description: User created
        "400":
          description: Weak password
        "401":
          description: Unauthorized
        "403":
          description: Admin privileges required
        "409":
          description: Email already registered
  /auth/admin/user/{user_id}:
    get:
      summary: Get user by ID (admin)
      tags: [Auth Admin]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: user_id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User details
        "401":
          description: Unauthorized
        "403":
          description: Admin privileges required
        "404":
          description: User not found
    patch:
      summary: Update user (admin)
      tags: [Auth Admin]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: user_id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                user_name:
                  type: string
                email:
                  type: string
                  format: email
                password:
                  type: string
                date_of_birth:
                  type: string
                is_admin:
                  type: boolean
      responses:
        "200":
          description: Updated user
        "400":
          description: Invalid payload
        "401":
          description: Unauthorized
        "403":
          description: Admin privileges required
        "404":
          description: User not found
        "409":
          description: Email already registered
    delete:
      summary: Delete user (admin)
      tags: [Auth Admin]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: user_id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [reason_code]
              properties:
                reason_code:
                  type: string
                  enum: [test_purpose, user_request, spam_abuse, security, other]
                reason_detail:
                  type: string
      responses:
        "200":
          description: User deleted
        "401":
          description: Unauthorized
        "403":
          description: Admin privileges required
        "404":
          description: User not found
  /auth/admin/user-deletion-logs:
    get:
      summary: List deletion audit logs (admin)
      tags: [Auth Admin]
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: page
          schema:
            type: integer
        - in: query
          name: page_size
          schema:
            type: integer
      responses:
        "200":
          description: Deletion logs
        "401":
          description: Unauthorized
        "403":
          description: Admin privileges required
  /auth/admin/access-requests:
    get:
      summary: List pending access requests (admin)
      tags: [Auth Admin]
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: page
          schema:
            type: integer
        - in: query
          name: page_size
          schema:
            type: integer
      responses:
        "200":
          description: Pending access requests
        "401":
          description: Unauthorized
        "403":
          description: Admin privileges required
  /auth/admin/access-requests/{user_id}/approve:
    post:
      summary: Approve access request (admin)
      tags: [Auth Admin]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: user_id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User approved
        "401":
          description: Unauthorized
        "403":
          description: Admin privileges required
        "404":
          description: No pending request
  /auth/admin/access-requests/{user_id}/reject:
    delete:
      summary: Reject access request (admin)
      tags: [Auth Admin]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: user_id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User rejected
        "401":
          description: Unauthorized
        "403":
          description: Admin privileges required
        "404":
          description: No pending request

  # ───────────────── Detection ─────────────────
...